+++ init/pxe/pxe-server/dhcp/clean-client.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo | sudo tee /var/lib/dhcpd/dhcpd.leases